backend/requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
python-dotenv==1.0.0
SQLAlchemy==2.0.19
google-generativeai==0.8.5
gunicorn==20.1.0